+++ src/app/collective-rewards/rewards/components/Svg/RbtcSvg.tsx
