+++ src/app/collective-rewards/rewards/components/Svg/RifSvg.tsx
